refactor: extract reply helper
